rhine-mine: break up Main
